split api into useUser, useMongo hooks
