+++ site/gatsby-site/src/mongodb/useUser.js
@@ -1,0 +1,20 @@
+import { useEffect, useState } from 'react';
+import { getUser } from './authenticate';
+
+export const useUser = () => {
+  const [loading, setLoading] = useState(true);
+
+  const [user, setUser] = useState();
+
+  useEffect(() => {
+    getUser().then((res) => {
+      setUser(res.user);
+      setLoading(false);
+    });
+  }, []);
+
+  return {
+    loading,
+    user,
+  };
+};
